folder structure cleanup
